api/requirements-lambda.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
supabase==1.0.3
pillow==9.5.0
numpy<2.0
opencv-python-headless==4.8.0.74
aiohttp>=3.9.0
PyJWT==2.8.0
boto3==1.34.0
mangum==0.17.0
typing_extensions>=4.0.0
# Compatible PyTorch for Python 3.9
torch==2.0.1
torchvision==0.15.2
ultralytics==8.2.28 